+++ demo/app/app.module.ts
@@ -25,6 +25,7 @@ import { AppMaterialModule } from './app.material.module';
 import { DemoMapComponent } from './components/demo-map/demo-map.component';
 import { MaxValidatorDirective } from './shared/settings-form/max-validtor.directive';
 import { MinValidatorDirective } from './shared/settings-form/min-validator.directive';
+import { MapsLayerComponent } from './components/maps-layer/maps-layer.component';
 
 
 @NgModule({
@@ -45,6 +46,7 @@ import { MinValidatorDirective } from './shared/settings-form/min-validator.dire
     TracksDialogComponent,
     DrawOnMapComponent,
     DemoMapComponent,
+    MapsLayerComponent,
     MaxValidatorDirective,
     MinValidatorDirective,
   ],