+++ demo/app/components/demo-map/demo-map.component.ts
@@ -12,13 +12,13 @@ import { TracksLayerComponent } from '../tracks-layer/tracks-layer.component';
 })
 export class DemoMapComponent {
   @Input() tracksRealData: boolean;
-
+  
   @ViewChild('layer') tracksLayer: TracksLayerComponent;
   arcGisMapServerProvider = MapLayerProviderOptions.ArcGisMapServer;
-
+  
   constructor(private viewerConf: ViewerConfiguration, private appSettingsService: AppSettingsService) {
     viewerConf.viewerOptions = {
-      sceneMode: Cesium.SceneMode.COLUMBUS_VIEW,
+      // sceneMode: Cesium.SceneMode.COLUMBUS_VIEW,
       selectionIndicator: false,
       timeline: false,
       infoBox: false,
@@ -30,18 +30,18 @@ export class DemoMapComponent {
       navigationHelpButton: false,
       navigationInstructionsInitiallyVisible: false,
     };
-
+    
     viewerConf.viewerModifier = (viewer) => {
       viewer.screenSpaceEventHandler.removeInputAction(Cesium.ScreenSpaceEventType.LEFT_DOUBLE_CLICK);
       viewer.bottomContainer.remove();
       const screenSpaceCameraController = viewer.scene.screenSpaceCameraController;
-      screenSpaceCameraController.enableTilt = false;
-      screenSpaceCameraController.enableRotate = false;
+      // screenSpaceCameraController.enableTilt = false;
+      // screenSpaceCameraController.enableRotate = false;
     };
-
+    
     this.appSettingsService.showTracksLayer = true;
-	}
-
+  }
+  
   removeAll() {
     this.tracksLayer.removeAll();
   }