+++ src/models/map-layer-provider-options.enum.ts
@@ -1,6 +1,14 @@
 export enum MapLayerProviderOptions {
-    ArcGisMapServer,
-    WebMapTileService,
-    WebMapService,
-    OFFLINE
+  ArcGisMapServer = Cesium.ArcGisMapServerImageryProvider ,
+  WebMapTileService = Cesium.WebMapTileServiceImageryProvider,
+  MapTileService =  Cesium.createTileMapServiceImageryProvider,
+  WebMapService = Cesium.WebMapServiceImageryProvider,
+  SingleTileImagery = Cesium.SingleTileImageryProvider,
+  OpenStreetMap =  Cesium.createOpenStreetMapImageryProvider,
+  BingMaps = Cesium.BingMapsImageryProvider,
+  GoogleEarthEnterpriseMaps = Cesium.GoogleEarthEnterpriseMapsProvider,
+  MapBox = Cesium.MapboxImageryProvider,
+  UrlTemplateImagery = Cesium. UrlTemplateImageryProvider,
+  OFFLINE = null,
 }
+