+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * AcEntity id filed changed to string from number
 * AcNotification id filed changed to string from number
 
+## 0.0.13
+### Features
+* ac-map-layer-provider - Support all cesium imagery providers
+* Expose cesium viewer through ViewesManagerService and `ac-map`
 
 ## 0.0.14
 ### Breaking changes